+++ mipi_cam/src/x3/sensor_ov5647_config.c
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "sensor_f37_config.h"
+#include "sensor_ov5647_config.h"
 
-MIPI_SENSOR_INFO_S SENSOR_1LANE_F37_30FPS_10BIT_LINEAR_INFO =
-{
+MIPI_SENSOR_INFO_S SENSOR_2LANE_OV5647_30FPS_10BIT_LINEAR_INFO = {
 	.deseEnable = 0, // 该 sensor 是否有 serdes（串解器）
 	.inputMode = INPUT_MODE_MIPI, // sensor 接入方式,mipi还是dvp
 	.sensorInfo = {
@@ -25,37 +24,36 @@ MIPI_SENSOR_INFO_S SENSOR_1LANE_F37_30FPS_10BIT_LINEAR_INFO =
 		.bus_num = 2, // 总线号，根据具体板子硬件原理图确定 , 不配置默认 i2c5
 		.fps = 30, // 帧率，用来选择使用哪一组帧率的sensor参数
 		.resolution = 1080, // sensor 行宽, 必须要和mipi属性配置一致
-		.sensor_addr = 0x40, // sensor i2c 设备地址
+		.sensor_addr = 0x36, // sensor i2c 设备地址
 		.entry_index = 0, // sensor 使用的 mipi 索引, 0~3，对应mipi host的序号
 		.sensor_mode = NORMAL_M, // sensor 工作模式， 1 是 normal,2 是dol2,3 是 dol3
-		.reg_width = 8, // sensor 寄存器地址宽度
-		.sensor_name = "f37", // sensor的名字，在libcam.so中会根据这个名字组合出 libf37.so
+		.reg_width = 16, // sensor 寄存器地址宽度
+		.sensor_name = "ov5647", // sensor的名字，在libcam.so中会根据这个名字组合出 libf37.so
 							// 这个库文件，然后调用dlopen运行时打开sensor库， libf37.so 这个库的生成
 							// 请参考sensor点亮文档，以f37 为例是 f37_utility.c f37_utility.h
 							// 请放到sdk包的hbre/camera/utility/sensor目录下make生成libf37.so文件
 	}
 };
 
-MIPI_ATTR_S MIPI_1LANE_SENSOR_F37_30FPS_10BIT_LINEAR_ATTR =
-{
+MIPI_ATTR_S MIPI_2LANE_SENSOR_OV5647_30FPS_10BIT_LINEAR_ATTR = {
     .mipi_host_cfg = {
-        .lane = 1, // 硬件上sensor用了几个mipi数据lane，f37用的1 lane, os8a10 4K的用了4 lane
+        .lane = 2, // 硬件上sensor用了几个mipi数据lane，f37用的1 lane, os8a10 4K的用了4 lane
         .datatype = 0x2b, // sensor的输出数据类型，请参考 《X3J3平台AIOT媒体系统接口手册.pdf》第3.5.36节 DATA TYPE
         .mclk = 2400, // mipi 模块主时钟，目前默认是24MHz
-        .mipiclk = 864, // sensor 输出总的 mipibit rate, 单位Mbits/秒
+        .mipiclk = 1000, // sensor 输出总的 mipibit rate, 单位Mbits/秒
         .fps = 30, // sensor 输出实际帧率, 和 MIPI_SENSOR_INFO_S 是什么关系？
         .width = 1920, // sensor实际输出的分辨率，一般根据sensor供应商的配置参数配置就行
         .height = 1080,
-        .linelenth = 2560, // sensor 输出带 blanking 的总行长， 一般根据sensor供应商的配置参数配置就行
-        .framelenth = 1125, // sensor 输出带 blanking的总行数
-        .settle = 20, // sensor 输出实际Ttx-zero + Ttx-prepare时间（ clk 为单位）, 一般0-100之间尝试
+        .linelenth = 2416, // sensor 输出带 blanking 的总行长， 一般根据sensor供应商的配置参数配置就行
+        .framelenth = 1104, // sensor 输出带 blanking的总行数
+        .settle = 30, // sensor 输出实际Ttx-zero + Ttx-prepare时间（ clk 为单位）, 一般0-100之间尝试
 		.channel_num = 1, // 使用虚通道的个数, linear 用一个，dol2 用 2 个， dol3 用3个
 		.channel_sel = {0} // 保存每个虚通道的值，在调用HB_VIN_SetDevVCNumber接口时用第一个，调用HB_VIN_AddDevVCNumber用后面的2 3 4的配置
     },
     .dev_enable = 0 // mipi dev 是否使能， 1是使能， 0 是关闭, 只用开启mipi bypass的开启
 };
 
-VIN_DEV_ATTR_S DEV_ATTR_F37_LINEAR_BASE = {
+VIN_DEV_ATTR_S DEV_ATTR_OV5647_LINEAR_BASE = {
     .stSize = { // 输入的数据
 		.format = 0, // 像素格式， format 为 0 代表是raw, 根据 pixel_lenght 来表示究竟是 raw8, raw12 还是 raw16 。
 		.width = 1920, // 数据宽
@@ -103,7 +101,7 @@ VIN_DEV_ATTR_S DEV_ATTR_F37_LINEAR_BASE = {
     }
 };
 
-VIN_PIPE_ATTR_S PIPE_ATTR_F37_LINEAR_BASE = {
+VIN_PIPE_ATTR_S PIPE_ATTR_OV5647_LINEAR_BASE = {
 	.ddrOutBufNum = 6, // ISP->GDC ddr的buf数量，最少可以设置成 2
 	.snsMode = SENSOR_NORMAL_MODE, // sensor 工作模式，要和MIPI_SENSOR_INFO_S配置的sensor_mode一致
 	.stSize = { // 数据格式同 VIN_DEV_SIZE_S
@@ -112,17 +110,17 @@ VIN_PIPE_ATTR_S PIPE_ATTR_F37_LINEAR_BASE = {
 		.height = 1080,
 	},
 	.cfaPattern = PIPE_BAYER_BGGR, // 数据格式布局(sensor输出的图像格式是RGGB还是BGGR),和sensor保持一致，查sensor数据手册可知
-	.temperMode = 2, // 时序降噪模式，2代表2帧融合，3代表3帧融合
+	.temperMode = 0, // 时序降噪模式，2代表2帧融合，3代表3帧融合
 	.ispBypassEn = 0, // 是否使能 isp 的 bypass, sensor必须支持输出yuv
 	.ispAlgoState = 1, // 是否启动 3a 算法库 ,1 是启动， 0 是关闭
 	.bitwidth = 10, // raw图位宽，有效值 8 、 10 、 12 、14 、 16 、 20
 	.calib = { // 是否开启 sensor 矫正数据加载，1 是开启，0 是关闭。ISP使用这些数据
 		.mode = 1, // 是否开启 sensor 矫正数据加载
-		.lname = "libjxf37_linear.so", // 对应使用的校准库
+		.lname = "libov5647_linear.so", // 对应使用的校准库
 	}
 };
 
-VIN_DIS_ATTR_S DIS_ATTR_F37_LINEAR_BASE = {
+VIN_DIS_ATTR_S DIS_ATTR_OV5647_LINEAR_BASE = {
 	.picSize = { // 输入数据宽高
 		.pic_w = 1919, // 需要设置比接入尺寸 -1 的 size, 如果 ISP 输出1920 , 则这里设置 1919
 		.pic_h = 1079, // 需要设置比接入尺寸 -1 的 size
@@ -144,7 +142,7 @@ VIN_DIS_ATTR_S DIS_ATTR_F37_LINEAR_BASE = {
 	.disBufNum = 8,
 };
 
-VIN_LDC_ATTR_S LDC_ATTR_F37_LINEAR_BASE = {
+VIN_LDC_ATTR_S LDC_ATTR_OV5647_LINEAR_BASE = {
   .ldcEnable = 0, // LDC 是否使能
   .ldcPath = { // 输出类型
 	 .rg_y_only = 0, // ???
@@ -179,223 +177,4 @@ VIN_LDC_ATTR_S LDC_ATTR_F37_LINEAR_BASE = {
 	 .rg_start = 0,
 	 .rg_length = 1079,
   }
-};
-
-MIPI_SENSOR_INFO_S SENSOR_1LANE_F37_30FPS_10BIT_DOL2_INFO =
-{
-	.deseEnable = 0,
-	.inputMode = INPUT_MODE_MIPI,
-	.sensorInfo = {
-		.port = 0,
-		.dev_port = 0,
-		.bus_type = 0,
-		.bus_num = 2,
-		.fps = 30,
-		.resolution = 1080,
-		.sensor_addr = 0x40,
-		.entry_index = 0,
-		.sensor_mode = DOL2_M,
-		.reg_width = 8,
-		.sensor_name = "f37",
-	}
-};
-
-MIPI_ATTR_S MIPI_1LANE_SENSOR_F37_30FPS_10BIT_DOL2_ATTR =
-{
-	.mipi_host_cfg = {
-	    .lane = 1,
-	    .datatype = 0x2b,
-	    .mclk = 2400,
-	    .mipiclk = 864,
-	    .fps = 30,
-	    .width = 1920,
-	    .height = 1080,
-	    .linelenth = 2560,
-	    .framelenth = 1125,
-	    .settle = 20,
-		.channel_num= 2,
-		.channel_sel = {0, 1}
-	},
-	.dev_enable = 0
-};
-
-VIN_DEV_ATTR_S DEV_ATTR_F37_DOL2_BASE = {
-    .stSize = {
-		.format = 0,
-		.width = 1920,
-		.height = 1080,
-		.pix_length = 1
-    },
-    .mipiAttr = {
-        .enable = 1,
-        .ipi_channels = 2,
-        .ipi_mode = 2,
-        .enable_mux_out = 1,
-        .enable_frame_id = 1,
-        .enable_bypass = 0,
-        .enable_line_shift = 0,
-        .enable_id_decoder = 0,
-        .set_init_frame_id = 1,
-        .set_line_shift_count = 0,
-        .set_bypass_channels = 1,
-    },
-    .DdrIspAttr = {
-        .buf_num = 4,
-        .raw_feedback_en = 0,
-        .data = {
-            .format = 0,
-            .width = 1920,
-            .height = 1080,
-            .pix_length = 1,
-        }
-    },
-    .outDdrAttr = {
-        .stride = 2400,
-        .buffer_num = 8,
-    },
-    .outIspAttr = {
-        .dol_exp_num = 2,
-        .enable_dgain = 0,
-        .set_dgain_short = 0,
-        .set_dgain_medium = 0,
-        .set_dgain_long = 0,
-        .vc_short_seq = 0,
-        .vc_medium_seq = 0,
-        .vc_long_seq = 1,
-    }
-};
-
-VIN_PIPE_ATTR_S PIPE_ATTR_F37_DOL2_BASE = {
-	.ddrOutBufNum = 6,
-	.snsMode = SENSOR_DOL2_MODE,
-	.stSize = {
-		.format = 0,
-		.width = 1920,
-		.height = 1080,
-	},
-	.snsMode = 1,
-	.cfaPattern = PIPE_BAYER_BGGR,
-	.temperMode = 2,
-	.ispBypassEn = 0,
-	.ispAlgoState = 1,
-	.bitwidth = 10,
-	.calib = {
-		.mode = 1,
-		.lname = "libjxf37_hdr.so",
-	}
-};
-
-VIN_DIS_ATTR_S DIS_ATTR_F37_DOL2_BASE = {
-	.picSize = {
-		.pic_w = 1919,
-		.pic_h = 1079,
-	},
-	.disPath = {
-		.rg_dis_enable = 0,
-		.rg_dis_path_sel = 1,
-	},
-	.disHratio = 65536,
-	.disVratio = 65536,
-	.xCrop = {
-		.rg_dis_start = 0,
-		.rg_dis_end = 1919,
-	},
-	.yCrop = {
-		.rg_dis_start = 0,
-		.rg_dis_end = 1079,
-	},
-	.disBufNum = 8,
-};
-
-VIN_LDC_ATTR_S LDC_ATTR_F37_DOL2_BASE = {
-  .ldcEnable = 0,
-  .ldcPath = {
-	 .rg_y_only = 0,
-	 .rg_uv_mode = 0,
-	 .rg_uv_interpo = 0,
-	 .rg_h_blank_cyc = 32,
-  },
-  .yStartAddr = 524288,
-  .cStartAddr = 786432,
-  .picSize = {
-	 .pic_w = 1919,
-	 .pic_h = 1079,
-  },
-  .lineBuf = 99,
-  .xParam = {
-	 .rg_algo_param_b = 1,
-	 .rg_algo_param_a = 1,
-  },
-  .yParam = {
-	 .rg_algo_param_b = 1,
-	 .rg_algo_param_a = 1,
-  },
-  .offShift = {
-	 .rg_center_xoff = 0,
-	 .rg_center_yoff = 0,
-  },
-  .xWoi = {
-	 .rg_start = 0,
-	 .rg_length = 1919,
-  },
-  .yWoi = {
-	 .rg_start = 0,
-	 .rg_length = 1079,
-  }
-};
-
-VIN_DEV_ATTR_S DEV_ATTR_F37_LINEAR_FEEDBACK = {
-	.stSize = {
-		0,	  /*format*/
-		1920, /*width*/
-		1080, /*height*/
-		1	  /*pix_length*/
-	},
-	.mipiAttr = {
-		.enable = 1,
-		.ipi_channels = 1,
-		.enable_mux_out = 1,
-		.enable_frame_id = 1,
-		.enable_bypass = 0,
-		.enable_line_shift = 0,
-		.enable_id_decoder = 0,
-		.set_init_frame_id = 1,
-		.set_line_shift_count = 0,
-		.set_bypass_channels = 1,
-	},
-	.DdrIspAttr = {
-		.buf_num = 4,
-		.raw_feedback_en = 1,
-		.data = {
-			.format = 0,
-			.width = 1920,
-			.height = 1080,
-			.pix_length = 1,
-		}},
-	.outDdrAttr = {
-		.stride = 2400,
-		.buffer_num = 8,
-	},
-	.outIspAttr = {
-		.dol_exp_num = 1,
-		.enable_dgain = 0,
-		.set_dgain_short = 0,
-		.set_dgain_medium = 0,
-		.set_dgain_long = 0,
-		.vc_short_seq = 0,
-		.vc_medium_seq = 0,
-		.vc_long_seq = 0,
-}};
-
-VIN_DEV_ATTR_EX_S DEV_ATTR_F37_MD_BASE = {
-	.path_sel = 0,
-	.roi_top = 0,
-	.roi_left = 0,
-	.roi_width = 1280,
-	.roi_height = 640,
-	.grid_step = 128,
-	.grid_tolerance = 10,
-	.threshold = 10,
-	.weight_decay = 128,
-	.precision = 0
 };